добавленые репозиторий и сервис для филиала; тип данных id изменен на long; мелкие правки

+++ src/main/java/ru/alexanderrogachev/staffer/repositories/RequestRepository.java
@@ -5,7 +5,7 @@ import ru.alexanderrogachev.staffer.models.Request;
 
 import java.util.List;
 
-public interface RequestRepository extends JpaRepository<Request, Integer> {
+public interface RequestRepository extends JpaRepository<Request, Long> {
 
     List<Request> findByShopName(String shopName);
